refactor(mcp): rename runtime_pi/discovery_pi to cubepi_runtime/cubepi_discovery (M6.5)
The langgraph-era cubebox/mcp/{runtime,discovery}.py are still in use by
admin / OAuth tool-refresh paths (services/mcp*.py, oauth/callback.py).
To avoid colliding with them, the cubepi-flavoured agent-runtime modules
are renamed with a cubepi_ prefix instead of stripping the _pi suffix:

- cubebox/mcp/runtime_pi.py     → cubebox/mcp/cubepi_runtime.py
- cubebox/mcp/discovery_pi.py   → cubebox/mcp/cubepi_discovery.py
- tests/unit/test_mcp_runtime_pi.py → test_mcp_cubepi_runtime.py

Function/class names (load_workspace_mcp_tools_for_cubepi,
CubepiMCPServerSpec, etc.) keep their cubepi_ infix to remain consistent
with the module path and stay distinguishable from the admin runtime.

Imports in run_manager.py + the test file are updated to match.
