+++ backend/cubebox/mcp/cubepi_discovery.py
